make location types serializable
