--- avs-api/src/test/java/org/example/age/avs/api/SiteLocationTest.java
@@ -1,16 +1,0 @@
-package org.example.age.avs.api;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import org.junit.jupiter.api.Test;
-
-public final class SiteLocationTest {
-
-    @Test
-    public void urls() {
-        SiteLocation location =
-                SiteLocation.builder("localhost", 80).redirectPath("verify").build();
-        assertThat(location.ageCertificateUrl().toString()).isEqualTo("http://localhost/api/age-certificate");
-        assertThat(location.redirectUrl().toString()).isEqualTo("http://localhost/verify");
-    }
-}
