+++ common/src/main/java/org/example/age/common/avs/verification/internal/Verification.java
@@ -1,6 +1,6 @@
 package org.example.age.common.avs.verification.internal;
 
-import org.example.age.avs.api.SiteLocation;
+import org.example.age.common.service.config.SiteLocation;
 import org.example.age.data.certificate.AgeCertificate;
 import org.immutables.value.Value;
