+++ common/src/main/java/org/example/age/common/avs/verification/internal/VerificationManagerImpl.java
@@ -9,7 +9,6 @@ import javax.inject.Named;
 import javax.inject.Provider;
 import javax.inject.Singleton;
 import org.example.age.api.HttpOptional;
-import org.example.age.avs.api.SiteLocation;
 import org.example.age.common.api.data.AuthMatchData;
 import org.example.age.common.api.data.AuthMatchDataExtractor;
 import org.example.age.common.avs.config.RegisteredSiteConfig;
@@ -18,6 +17,7 @@ import org.example.age.common.avs.store.VerifiedUserStore;
 import org.example.age.common.base.store.PendingStore;
 import org.example.age.common.base.store.PendingStoreFactory;
 import org.example.age.common.base.utils.internal.PendingStoreUtils;
+import org.example.age.common.service.config.SiteLocation;
 import org.example.age.common.service.data.internal.AuthMatchDataEncryptor;
 import org.example.age.data.certificate.AgeCertificate;
 import org.example.age.data.certificate.VerificationRequest;
