+++ common/src/test/java/org/example/age/common/avs/verification/VerificationManagerTest.java
@@ -14,7 +14,6 @@ import javax.inject.Named;
 import javax.inject.Singleton;
 import org.assertj.core.data.Offset;
 import org.example.age.api.HttpOptional;
-import org.example.age.avs.api.SiteLocation;
 import org.example.age.common.avs.config.RegisteredSiteConfig;
 import org.example.age.common.avs.store.InMemoryRegisteredSiteConfigStoreModule;
 import org.example.age.common.avs.store.InMemoryVerifiedUserStoreModule;
@@ -24,6 +23,7 @@ import org.example.age.common.avs.verification.internal.Verification;
 import org.example.age.common.avs.verification.internal.VerificationManager;
 import org.example.age.common.avs.verification.internal.VerificationManagerModule;
 import org.example.age.common.base.store.InMemoryPendingStoreFactoryModule;
+import org.example.age.common.service.config.SiteLocation;
 import org.example.age.common.service.data.DisabledAuthMatchDataExtractorModule;
 import org.example.age.data.certificate.VerificationSession;
 import org.example.age.data.crypto.SecureId;
