+++ site-service/src/main/java/org/example/age/site/service/config/internal/SiteConfigurerModule.java
@@ -5,8 +5,8 @@ import dagger.Provides;
 import java.security.PublicKey;
 import java.time.Duration;
 import javax.inject.Named;
+import org.example.age.common.service.config.AvsLocation;
 import org.example.age.data.crypto.SecureId;
-import org.example.age.site.api.AvsLocation;
 import org.example.age.site.service.config.SiteConfig;
 
 /**
