+++ site-service/src/test/java/org/example/age/site/service/SiteServiceTest.java
@@ -16,13 +16,13 @@ import java.util.Map;
 import javax.inject.Named;
 import javax.inject.Singleton;
 import okhttp3.Response;
-import org.example.age.avs.api.SiteLocation;
+import org.example.age.common.service.config.AvsLocation;
+import org.example.age.common.service.config.SiteLocation;
 import org.example.age.common.service.data.DisabledAuthMatchDataExtractorModule;
 import org.example.age.common.service.store.InMemoryPendingStoreFactoryModule;
 import org.example.age.data.certificate.VerificationSession;
 import org.example.age.data.crypto.SecureId;
 import org.example.age.data.crypto.SigningKeys;
-import org.example.age.site.api.AvsLocation;
 import org.example.age.site.service.config.SiteConfig;
 import org.example.age.site.service.store.InMemoryVerificationStoreModule;
 import org.example.age.test.avs.service.FakeAvsServiceModule;
