+++ site-service/src/testFixtures/java/org/example/age/test/avs/service/FakeAvsService.java
@@ -14,8 +14,8 @@ import org.example.age.api.Dispatcher;
 import org.example.age.api.JsonSender;
 import org.example.age.api.StatusCodeSender;
 import org.example.age.avs.api.AvsApi;
-import org.example.age.avs.api.SiteLocation;
 import org.example.age.common.api.data.AuthMatchData;
+import org.example.age.common.service.config.SiteLocation;
 import org.example.age.common.service.data.internal.AuthMatchDataEncryptor;
 import org.example.age.data.certificate.AgeCertificate;
 import org.example.age.data.certificate.SignedAgeCertificate;
